fixed data type

+++ app/src/main/java/com/example/app/model/GetAllExerciseRecordResponse.kt
@@ -2,17 +2,17 @@ package com.example.app.model
 
 data class ExerciseRecord(
     val ID: Int,
-    val avgHeartRate: Int,
-    val bloodOxygenLevel: Int,
+    val avgHeartRate: Float,
+    val bloodOxygenLevel: Float,
     val caloriesBurned: Float,
     val createdAt: String,
     val distance: Float,
     val duration: Int,
     val endTime: String,
     val exerciseType: String,
-    val highBloodPressure: Int,
+    val highBloodPressure: Float,
     val locationPath: String,
-    val lowBloodPressure: Int,
+    val lowBloodPressure: Float,
     val startTime: String,
     val stepsCount: Int,
     val updatedAt: String
